+++ titanium/__init__.py
@@ -1,5 +1,5 @@
 APP_NAME = "Titanium APK Builder"
-VERSION = "10.0.0-dev.1"
+VERSION = "10.0.0-dev.2"
 ANDROID_API = 36
 BUILD_TOOLS = "36.0.0"
 AGP_VERSION = "9.4.0"